refactor unnessasary model
