+++ app/src/main/java/hu/gyuriczaadam/sprintformteszt/data/local/data_source/TransactionsDatabase.kt
@@ -3,12 +3,12 @@ package hu.gyuriczaadam.sprintformteszt.data.local.data_source
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import hu.gyuriczaadam.sprintformteszt.data.local.daos.TransactionDao
-import hu.gyuriczaadam.sprintformteszt.domain.model.local.TransactionListEntity
-import hu.gyuriczaadam.sprintformteszt.domain.model.remote.TransactionListModel
+import hu.gyuriczaadam.sprintformteszt.data.local.entities.TransactionListEntity
+
 
 @Database(
     entities = [TransactionListEntity::class],
-    version = 1
+    version = 2
 )
 abstract class TransactionsDatabase :RoomDatabase(){
     abstract val transactionDao:TransactionDao
