+++ app/src/main/java/hu/gyuriczaadam/sprintformteszt/domain/repositories/TransactionRepository.kt
@@ -1,8 +1,8 @@
 package hu.gyuriczaadam.sprintformteszt.domain.repositories
 
+import hu.gyuriczaadam.sprintformteszt.data.local.entities.TransactionListEntity
 import hu.gyuriczaadam.sprintformteszt.data.remote.transaction_dto.TransactionDto
 import hu.gyuriczaadam.sprintformteszt.domain.model.TransactionItem
-import hu.gyuriczaadam.sprintformteszt.domain.model.local.TransactionListEntity
 import kotlinx.coroutines.flow.Flow
 
 interface TransactionRepository {
