+++ app/src/main/java/hu/gyuriczaadam/sprintformteszt/domain/use_case/GetTransactionsFromApiUseCase.kt
@@ -1,16 +1,12 @@
 package hu.gyuriczaadam.sprintformteszt.domain.use_case
 
-import hu.gyuriczaadam.sprintformteszt.domain.model.remote.TransactionListModel
 import hu.gyuriczaadam.sprintformteszt.domain.repositories.TransactionRepository
 import hu.gyuriczaadam.sprintformteszt.util.Resource
 import hu.gyuriczaadam.sprintformteszt.util.UIText
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import okio.IOException
 import hu.gyuriczaadam.sprintformteszt.R
-import hu.gyuriczaadam.sprintformteszt.data.remote.transaction_dto.toTransactionListModel
-import hu.gyuriczaadam.sprintformteszt.domain.model.local.TransactionListEntity
-import hu.gyuriczaadam.sprintformteszt.domain.model.remote.toTransactionListEntity
+import hu.gyuriczaadam.sprintformteszt.data.remote.transaction_dto.toTransactionEntity
 import retrofit2.HttpException
 import javax.inject.Inject
 
@@ -19,8 +15,8 @@ class GetTransactionsFromApiUseCase @Inject constructor(private val repository: 
     operator fun invoke() = flow {
         try {
             emit(Resource.Loading())
-            val transactions = repository.getTransactionsFromApi().map {it.toTransactionListModel()}
-            val transactionsList= transactions.map { repository.insertTransaction(it.toTransactionListEntity()) }
+            val transactions = repository.getTransactionsFromApi().map {it.toTransactionEntity()}
+            val transactionsList= transactions.map {repository.insertTransaction(it)}
             emit(Resource.Success(transactionsList))
         }catch (e:HttpException){
             emit(Resource.Error(UIText.DynamicString(e.localizedMessage)))
